LayoutTests/imported/w3c/web-platform-tests/css/css-values/if-media-invalidation-expected.txt



FAIL CSS Values and Units Test: if() media() condition invalidation assert_equals: --actual before resize expected "true_value" but got "if(media((height < 100px) or ((height >= 200px) and (height < 300px))): true_value; else: false_value;)"
